// src/services/models/ClientModel.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ClientModel = {
    id?: string | null;
    companyName?: string | null;
    name?: string | null;
    companyEmail?: string | null;
    email?: string | null;
    companyPhoneNumber?: string | null;
    phoneNumber?: string | null;
    companyAddress?: string | null;
    address?: string | null;
};